pyproject.toml
[tool.poetry]
name = "nonebot-plugin-sentry"
version = "1.0.1"
description = "Push your bot errors to Sentry.io"
license = "MIT"
authors = ["yanyongyu <yyy@nonebot.dev>"]
readme = "README.md"
homepage = "https://github.com/cscs181/QQ-GitHub-Bot/tree/master/src/plugins/nonebot_plugin_sentry"
repository = "https://github.com/cscs181/QQ-GitHub-Bot/tree/master/src/plugins/nonebot_plugin_sentry"
documentation = "https://github.com/cscs181/QQ-GitHub-Bot/tree/master/src/plugins/nonebot_plugin_sentry#readme"
keywords = ["nonebot", "nonebot2", "sentry"]

[tool.poetry.dependencies]
python = "^3.8"
nonebot2 = "^2.2.0"
pydantic = ">=1.10.0,<3.0.0,!=2.5.0,!=2.5.1"
sentry-sdk = { version = "^1.23.0", extras = ["loguru"] }

[tool.poetry.group.dev.dependencies]
ruff = "^0.2.0"
isort = "^5.10.1"
black = "^24.0.0"
nonemoji = "^0.1.3"
pre-commit = "^3.3.0"

[tool.black]
line-length = 88
include = '\.pyi?$'
extend-exclude = '''
'''

[tool.isort]
profile = "black"
line_length = 88
length_sort = true
skip_gitignore = true
force_sort_within_sections = true
extra_standard_library = ["typing_extensions"]

[tool.ruff]
line-length = 88
target-version = "py38"

[tool.ruff.lint]
select = ["E", "W", "F", "UP", "C", "T", "Q"]
ignore = ["E402", "F403", "F405", "C901", "UP037"]

[tool.pyright]
pythonPlatform = "All"
pythonVersion = "3.8"
defineConstant = { PYDANTIC_V2 = true }

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
